[REFACTORING] Moved more files around, reduced size of main again
