+++ src/main/java/life/qbic/model/dataLoading/QbicDataDownloader.java
@@ -1,4 +1,4 @@
-package life.qbic.dataLoading;
+package life.qbic.model.dataLoading;
 
 import ch.ethz.sis.openbis.generic.asapi.v3.IApplicationServerApi;
 import ch.ethz.sis.openbis.generic.asapi.v3.dto.dataset.DataSet;
@@ -10,6 +10,7 @@ import ch.ethz.sis.openbis.generic.dssapi.v3.dto.datasetfile.download.DataSetFil
 import ch.ethz.sis.openbis.generic.dssapi.v3.dto.datasetfile.id.DataSetFilePermId;
 import ch.ethz.sis.openbis.generic.dssapi.v3.dto.datasetfile.id.IDataSetFileId;
 import ch.systemsx.cisd.common.spring.HttpInvokerUtils;
+import life.qbic.io.commandline.PostmanCommandLineOptions;
 import life.qbic.util.ProgressBar;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -18,7 +19,7 @@ import java.io.*;
 import java.util.*;
 
 
-public class QbicDataLoader {
+public class QbicDataDownloader {
 
     private String user;
 
@@ -28,7 +29,7 @@ public class QbicDataLoader {
 
     private IDataStoreServerApi dataStoreServer;
 
-    private final static Logger LOG = LogManager.getLogger(QbicDataLoader.class);
+    private final static Logger LOG = LogManager.getLogger(QbicDataDownloader.class);
 
     private String sessionToken;
 
@@ -45,9 +46,9 @@ public class QbicDataLoader {
      * @param password The openBis password
      * @param bufferSize The buffer size for the InputStream reader
      */
-    public QbicDataLoader(String AppServerUri, String DataServerUri,
-                                         String user, String password,
-                                         int bufferSize, String filterType){
+    public QbicDataDownloader(String AppServerUri, String DataServerUri,
+                              String user, String password,
+                              int bufferSize, String filterType){
         this.defaultBufferSize = bufferSize;
         this.filterType = filterType;
 
@@ -76,7 +77,7 @@ public class QbicDataLoader {
      * @param password The openBIS user's password
      * @return QBiCDataLoader instance
      */
-    public QbicDataLoader setCredentials(String user, String password) {
+    public QbicDataDownloader setCredentials(String user, String password) {
         this.user = user;
         this.password = password;
         return this;
@@ -187,24 +188,107 @@ public class QbicDataLoader {
 
             }
         }
+
         return 0;
     }
 
+    /**
+     * Downloads the files that the user requested
+     * checks whether any filtering option (suffix or regex) has been passed and applies filtering if needed
+     *
+     * @param commandLineParameters
+     * @param qbicDataDownloader
+     * @throws IOException
+     */
+    public void downloadRequestedFilesOfDatasets(PostmanCommandLineOptions commandLineParameters, QbicDataDownloader qbicDataDownloader) throws IOException {
+        QbicDataFinder qbicDataFinder = new QbicDataFinder(applicationServer,
+                                                           dataStoreServer,
+                                                           sessionToken,
+                                                           filterType);
 
-    public IApplicationServerApi getApplicationServer() {
-        return applicationServer;
+        LOG.info(String.format("%s provided openBIS identifiers have been found: %s",
+                commandLineParameters.ids.size(), commandLineParameters.ids.toString()));
+
+        // a suffix was provided -> only download files which contain the suffix string
+        if (!commandLineParameters.suffixes.isEmpty()) {
+            for (String ident : commandLineParameters.ids) {
+                LOG.info(String.format("Downloading files for provided identifier %s", ident));
+                List<IDataSetFileId> foundSuffixFilteredIDs = qbicDataFinder.findAllSuffixFilteredIDs(ident, commandLineParameters.suffixes);
+
+                LOG.info(String.format("Number of files found: %s", foundSuffixFilteredIDs.size()));
+
+                downloadFilteredIDs(qbicDataDownloader, ident, foundSuffixFilteredIDs);
+            }
+            // a regex pattern was provided -> only download files which contain the regex pattern
+        } else if (!commandLineParameters.regexPatterns.isEmpty()) {
+            for (String ident : commandLineParameters.ids) {
+                LOG.info(String.format("Downloading files for provided identifier %s", ident));
+                List<IDataSetFileId> foundRegexFilteredIDs = qbicDataFinder.findAllRegexFilteredIDs(ident, commandLineParameters.regexPatterns);
+
+                LOG.info(String.format("Number of files found: %s", foundRegexFilteredIDs.size()));
+
+                downloadFilteredIDs(qbicDataDownloader, ident, foundRegexFilteredIDs);
+            }
+        } else {
+            // no suffix or regex was supplied -> download all datasets
+            for (String ident : commandLineParameters.ids) {
+                LOG.info(String.format("Downloading files for provided identifier %s", ident));
+                List<DataSet> foundDataSets = qbicDataFinder.findAllDatasetsRecursive(ident);
+
+                LOG.info(String.format("Number of data sets found: %s", foundDataSets.size()));
+
+                if (foundDataSets.size() > 0) {
+                    LOG.info("Initialize download ...");
+                    int datasetDownloadReturnCode = -1;
+                    try {
+                        datasetDownloadReturnCode = qbicDataDownloader.downloadDataset(foundDataSets);
+                    } catch (NullPointerException e) {
+                        LOG.error("Datasets were found by the application server, but could not be found on the datastore server for "
+                                + ident + "." + " Try to supply the correct datastore server using a config file!");
+                    }
+
+                    if (datasetDownloadReturnCode != 0) {
+                        LOG.error("Error while downloading dataset: " + ident);
+                    } else {
+                        LOG.info("Download successfully finished.");
+                    }
+
+                } else {
+                    LOG.info("Nothing to download.");
+                }
+            }
+        }
     }
 
-    public IDataStoreServerApi getDataStoreServer() {
-        return dataStoreServer;
+
+    /**
+     * downloads all IDs which were previously filtered by either suffixes or regexes
+     *
+     * @param qbicDataDownloader
+     * @param ident
+     * @param foundFilteredIDs
+     * @throws IOException
+     */
+    private static void downloadFilteredIDs(QbicDataDownloader qbicDataDownloader, String ident, List<IDataSetFileId> foundFilteredIDs) throws IOException {
+        if (foundFilteredIDs.size() > 0) {
+            LOG.info("Initialize download ...");
+            int filesDownloadReturnCode = -1;
+            try {
+                filesDownloadReturnCode = qbicDataDownloader.downloadFilesByID(foundFilteredIDs);
+            } catch (NullPointerException e) {
+                LOG.error("Datasets were found by the application server, but could not be found on the datastore server for "
+                        + ident + "." + " Try to supply the correct datastore server using a config file!");
+            }
+            if (filesDownloadReturnCode != 0) {
+                LOG.error("Error while downloading dataset: " + ident);
+            } else {
+                LOG.info("Download successfully finished");
+            }
+
+        } else {
+            LOG.info("Nothing to download.");
+        }
     }
 
-    public String getSessionToken() {
-        return sessionToken;
-    }
-
-    public String getFilterType() {
-        return filterType;
-    }
 }
